reogranize plugin code
